(feature) split views
