+++ _UI/_web_interface/app.py
@@ -18,13 +18,17 @@
 #
 # - coding: utf-8 -*-
 
-from app_layout import make_layout
+# isort: off
 from maindash import app
 
-app.layout = make_layout()
+# isort: on
+
+from views import main
+
+app.layout = main.layout
 
 # It is workaround for splitting callbacks in separate files (run callbacks after layout)
-from callbacks import display_page, update_daq_params  # noqa: F401
+from callbacks import display_page, main, update_daq_params  # noqa: F401
 
 if __name__ == "__main__":
     # Debug mode does not work when the data interface is set to shared-memory
