+++ _UI/_web_interface/callbacks/display_page.py
@@ -1,34 +1,13 @@
-import dash_core_components as dcc
-import dash_html_components as html
-
 # isort: off
-from maindash import app, spectrum_fig, waterfall_fig, webInterface_inst
+from maindash import app, webInterface_inst
 
 # isort: on
+
 from dash_devices.dependencies import Input, Output
 from kraken_web_config import generate_config_page_layout
-from kraken_web_doa import generate_doa_page_layout, plot_doa
+from kraken_web_doa import plot_doa
 from variables import doa_fig
-
-spectrum_page_layout = html.Div(
-    [
-        html.Div(
-            [
-                dcc.Graph(
-                    id="spectrum-graph",
-                    style={"width": "100%", "height": "45%"},
-                    figure=spectrum_fig,  # fig_dummy #spectrum_fig #fig_dummy
-                ),
-                dcc.Graph(
-                    id="waterfall-graph",
-                    style={"width": "100%", "height": "65%"},
-                    figure=waterfall_fig,  # waterfall fig remains unchanged always due to slow speed to update entire graph #fig_dummy #spectrum_fig #fig_dummy
-                ),
-            ],
-            style={"width": "100%", "height": "80vh"},
-        ),
-    ]
-)
+from views import generate_doa_page, spectrum_page
 
 
 @app.callback(
@@ -56,10 +35,10 @@ def display_page(pathname):
     elif pathname == "/spectrum":
         webInterface_inst.module_signal_processor.en_spectrum = True
         webInterface_inst.reset_spectrum_graph_flag = True
-        return [spectrum_page_layout, "header_inactive", "header_active", "header_inactive"]
+        return [spectrum_page.layout, "header_inactive", "header_active", "header_inactive"]
     elif pathname == "/doa":
         webInterface_inst.module_signal_processor.en_spectrum = False
         webInterface_inst.reset_doa_graph_flag = True
         plot_doa(app, webInterface_inst, doa_fig)
-        return [generate_doa_page_layout(webInterface_inst), "header_inactive", "header_inactive", "header_active"]
+        return [generate_doa_page.layout, "header_inactive", "header_inactive", "header_active"]
     return Output("dummy_output", "children", "")
