+++ _UI/_web_interface/views/tooltips.py
